[#11024] DocDB: Enable -Wshorten-64-to-32 flag
Summary:
This diff enables -Wshorten-64-to-32 flag and fixes all places where this warning appear in our code.
The following strategy was used:
1) Minimize amount of casts, i.e. changing used type is more preferable.
2) For places where such cast is necessary, for instance because of 3rd party library API, use newly introduced `narrow_cast`.
`narrow_cast` expects that casted values left unchanged. I.e. if we cast 64 bit number to 32 bit number it expects that actual value fits into 32 bits.

Test Plan: Jenkins

Reviewers: mbautin, mpolitov

Reviewed By: mpolitov

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D14653

+++ ent/src/yb/integration-tests/twodc_test_base.h
@@ -110,7 +110,7 @@ class TwoDCTestBase : public YBTest {
   CHECKED_STATUS DeleteUniverseReplication(
       const std::string& universe_id, YBClient* client, MiniCluster* cluster);
 
-  uint32_t NumProducerTabletsPolled(MiniCluster* cluster);
+  size_t NumProducerTabletsPolled(MiniCluster* cluster);
 
   CHECKED_STATUS CorrectlyPollingAllTablets(MiniCluster* cluster, uint32_t num_producer_tablets);
