+++ src/yb/client/client-internal.cc
@@ -1373,7 +1373,7 @@ CreateCDCStreamRpc::CreateCDCStreamRpc(YBClient* client,
       table_id_(table_id),
       options_(options) {
   req_.set_table_id(table_id_);
-  req_.mutable_options()->Reserve(options_.size());
+  req_.mutable_options()->Reserve(narrow_cast<int>(options_.size()));
   for (const auto& option : options_) {
     auto* op = req_.add_options();
     op->set_key(option.first);
