+++ src/yb/integration-tests/external_mini_cluster.h
@@ -86,10 +86,10 @@ using yb::consensus::ChangeConfigType;
 struct ExternalMiniClusterOptions {
 
   // Number of masters to start.
-  int num_masters = 1;
+  size_t num_masters = 1;
 
   // Number of TS to start.
-  int num_tablet_servers = 1;
+  size_t num_tablet_servers = 1;
 
   // Number of drives to use on TS.
   int num_drives = 1;
@@ -226,7 +226,7 @@ class ExternalMiniCluster : public MiniClusterBase {
   // Return the IP address that the tablet server with the given index will bind to.  If
   // options.bind_to_unique_loopback_addresses is false, this will be 127.0.0.1 Otherwise, it is
   // another IP in the local netblock.
-  std::string GetBindIpForTabletServer(int index) const;
+  std::string GetBindIpForTabletServer(size_t index) const;
 
   // Return a pointer to the running leader master. This may be NULL
   // if the cluster is not started.
@@ -298,16 +298,16 @@ class ExternalMiniCluster : public MiniClusterBase {
   CHECKED_STATUS WaitForLeaderCommitTermAdvance();
 
   // This API waits for the commit indices of all the master peers to reach the target index.
-  CHECKED_STATUS WaitForMastersToCommitUpTo(int target_index);
+  CHECKED_STATUS WaitForMastersToCommitUpTo(int64_t target_index);
 
   // If this cluster is configured for a single non-distributed master, return the single master or
   // NULL if the master is not started. Exits with a CHECK failure if there are multiple masters.
   ExternalMaster* master() const;
 
   // Return master at 'idx' or NULL if the master at 'idx' has not been started.
-  ExternalMaster* master(int idx) const;
+  ExternalMaster* master(size_t idx) const;
 
-  ExternalTabletServer* tablet_server(int idx) const;
+  ExternalTabletServer* tablet_server(size_t idx) const;
 
   // Return ExternalTabletServer given its UUID. If not found, returns NULL.
   ExternalTabletServer* tablet_server_by_uuid(const std::string& uuid) const;
@@ -328,11 +328,11 @@ class ExternalMiniCluster : public MiniClusterBase {
   // Get tablet server host.
   HostPort pgsql_hostport(int node_index) const;
 
-  int num_tablet_servers() const {
+  size_t num_tablet_servers() const {
     return tablet_servers_.size();
   }
 
-  int num_masters() const {
+  size_t num_masters() const {
     return masters_.size();
   }
 
@@ -383,7 +383,7 @@ class ExternalMiniCluster : public MiniClusterBase {
   // Wait until the number of registered tablet servers reaches the given count on at least one of
   // the running masters.  Returns Status::TimedOut if the desired count is not achieved with the
   // given timeout.
-  CHECKED_STATUS WaitForTabletServerCount(int count, const MonoDelta& timeout);
+  CHECKED_STATUS WaitForTabletServerCount(size_t count, const MonoDelta& timeout);
 
   // Runs gtest assertions that no servers have crashed.
   void AssertNoCrashes();
@@ -506,7 +506,7 @@ class ExternalMiniCluster : public MiniClusterBase {
 
   // This variable is incremented every time a new master is spawned (either in shell mode or create
   // mode). Avoids reusing an index of a killed/removed master. Useful for master side logging.
-  int add_new_master_at_;
+  size_t add_new_master_at_;
 
   std::vector<scoped_refptr<ExternalMaster> > masters_;
   std::vector<scoped_refptr<ExternalTabletServer> > tablet_servers_;
@@ -703,7 +703,7 @@ class ScopedResumeExternalDaemon {
 class ExternalMaster : public ExternalDaemon {
  public:
   ExternalMaster(
-    int master_index,
+    size_t master_index,
     rpc::Messenger* messenger,
     rpc::ProxyCache* proxy_cache,
     const std::string& exe,
@@ -731,7 +731,7 @@ class ExternalMaster : public ExternalDaemon {
 class ExternalTabletServer : public ExternalDaemon {
  public:
   ExternalTabletServer(
-      int tablet_server_index, rpc::Messenger* messenger, rpc::ProxyCache* proxy_cache,
+      size_t tablet_server_index, rpc::Messenger* messenger, rpc::ProxyCache* proxy_cache,
       const std::string& exe, const std::string& data_dir, uint16_t num_drives,
       std::string bind_host, uint16_t rpc_port, uint16_t http_port, uint16_t redis_rpc_port,
       uint16_t redis_http_port, uint16_t cql_rpc_port, uint16_t cql_http_port,
